refactor(android): use correct imports
