+++ apps/android/app/src/main/java/at/aau/serg/android/ui/screens/leaderboard/LeaderboardViewModel.kt
@@ -4,8 +4,8 @@ import at.aau.serg.android.core.network.RetrofitProvider
 import at.aau.serg.android.core.network.leaderboard.LeaderboardAPI
 import at.aau.serg.android.core.network.leaderboard.LeaderboardService
 import at.aau.serg.android.core.viewmodel.BaseViewModel
-import at.aau.serg.android.util.DispatcherProvider
-import at.aau.serg.android.util.DefaultDispatcherProvider
+import at.aau.serg.android.ui.util.DefaultDispatcherProvider
+import at.aau.serg.android.ui.util.DispatcherProvider
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import shared.models.LeaderboardEntry
